requirements.txt
aiosqlite
aiohttp
discord.py
python-dotenv